chore: remove leftover utils file
