+++ src/settings/svr_parameters.rs
@@ -1,6 +1,6 @@
 //! Support Vector Regression parameters
 
-pub use crate::utils::Kernel;
+pub use crate::utils::kernels::Kernel;
 
 /// Parameters for support vector regression
 #[derive(serde::Serialize, serde::Deserialize)]
